requirements.txt
openenv-core
fastapi
uvicorn
openai
